Fix MSG_MHF_SAVE_FAVORITE_QUEST parser

+++ network/mhfpacket/msg_mhf_save_favorite_quest.go
@@ -21,7 +21,7 @@ func (m *MsgMhfSaveFavoriteQuest) Opcode() network.PacketID {
 func (m *MsgMhfSaveFavoriteQuest) Parse(bf *byteframe.ByteFrame) error {
 	m.AckHandle = bf.ReadUint32()
 	m.DataSize = bf.ReadUint16()
-	m.Data = bf.ReadBytes(m.DataSize)
+	m.Data = bf.ReadBytes(uint(m.DataSize))
 }
 
 // Build builds a binary packet from the current data.
